shorten labels in score_bonuses.json
This is a fringe case, but with "tapettu" included there is a chance of the numbers overlapping each others. In the smaller margin, the context might not be understood, but in the wider perspective of the final score screen, that really shouldn't be an issue. Might change this back if there's negative feedback on this.